+++ lib/HAPServer.js
@@ -18,7 +18,7 @@ var hueLights = require('./util/hue').lights;
 var hueGroups = require('./util/hue').groups;
 var hueLight = require('./util/hue').light;
 var hueState = require('./util/hue').state;
-var hueDescription = require('./util/hue').description;
+var hueDescription = require('./util/ssdp').description;
 var ssdp = require('./util/ssdp');
 
 module.exports = {
