+++ lib/util/hue.js
@@ -1,13 +1,11 @@
 'use strict';
 
 var ip = require('ip');
-var port = require('./ssdp').getPort;
 
 module.exports = {
     lights: lights,
     groups: groups,
     light: light,
-    description: description,
     state: state
 };
 
@@ -18,10 +16,11 @@ function lights(accessories) {
 function state(url,requestData)
 {
 
-  console.log(url,requestData);
+
   var aid = parseInt(url.split("/")[4]);
 
 // { bri: 127 } = Dim
+// iid 10 is dimmer
 
   if ( requestData.on == false )
   {
@@ -78,57 +77,6 @@ function light(characteristics) {
       "uniqueid": "garbage",
       "swversion": "66012040"
   });
-}
-
-function description(accessories) {
-  let response = `<?xml version="1.0" encoding="UTF-8" ?>
-  <root xmlns="urn:schemas-upnp-org:device-1-0">
-  <specVersion>
-  <major>1</major>
-  <minor>0</minor>
-  </specVersion>
-  <URLBase>http://`+ip.address()+`:`+port()+`/</URLBase>
-  <device>
-  <deviceType>urn:schemas-upnp-org:device:Basic:1</deviceType>
-  <friendlyName>Philips hue (`+ip.address()+`)</friendlyName>
-  <manufacturer>Royal Philips Electronics</manufacturer>
-  <manufacturerURL>http://www.philips.com</manufacturerURL>
-  <modelDescription>Philips hue Personal Wireless Lighting</modelDescription>
-  <modelName>Philips hue bridge 2015</modelName>
-  <modelNumber>BSB002</modelNumber>
-  <modelURL>http://www.meethue.com</modelURL>
-  <serialNumber>E0CB4EFFFEE16375</serialNumber>
-  <UDN>uuid:2f402f80-da50-11e1-9b23-e0cb4ee16375</UDN>
-  <serviceList>
-  <service>
-  <serviceType>(null)</serviceType>
-  <serviceId>(null)</serviceId>
-  <controlURL>(null)</controlURL>
-  <eventSubURL>(null)</eventSubURL>
-  <SCPDURL>(null)</SCPDURL>
-  </service>
-  </serviceList>
-  <presentationURL>index.html</presentationURL>
-  <iconList>
-  <icon>
-  <mimetype>image/png</mimetype>
-  <height>48</height>
-  <width>48</width>
-  <depth>24</depth>
-  <url>hue_logo_0.png</url>
-  </icon>
-  <icon>
-  <mimetype>image/png</mimetype>
-  <height>120</height>
-  <width>120</width>
-  <depth>24</depth>
-  <url>hue_logo_3.png</url>
-  </icon>
-  </iconList>
-  </device>
-  </root>`
-
-  return(response);
 }
 
 function parseHbtoHue(accessories) {
