+++ lib/util/ssdp.js
@@ -8,6 +8,11 @@ const q = require('q');
 var ip = require('ip');
 var port;
 
+module.exports = {
+    startSsdpServer: startSsdpServer,
+    description: description
+};
+
 let udpServer;
 
 function getDiscoveryResponses() {
@@ -46,12 +51,7 @@ function stopDiscoveryServer() {
     return deferred.promise;
 }
 
-module.exports.getPort = function() {
-  return(port);
-}
-
-
-module.exports.startSsdpServer = function(_port) {
+function startSsdpServer(_port) {
 
     port = _port;
 
@@ -104,3 +104,54 @@ module.exports.startSsdpServer = function(_port) {
 };
 
 module.exports.stopDiscoveryServer = stopDiscoveryServer;
+
+function description(accessories) {
+  let response = `<?xml version="1.0" encoding="UTF-8" ?>
+  <root xmlns="urn:schemas-upnp-org:device-1-0">
+  <specVersion>
+  <major>1</major>
+  <minor>0</minor>
+  </specVersion>
+  <URLBase>http://`+ip.address()+`:`+port+`/</URLBase>
+  <device>
+  <deviceType>urn:schemas-upnp-org:device:Basic:1</deviceType>
+  <friendlyName>Philips hue (`+ip.address()+`)</friendlyName>
+  <manufacturer>Royal Philips Electronics</manufacturer>
+  <manufacturerURL>http://www.philips.com</manufacturerURL>
+  <modelDescription>Philips hue Personal Wireless Lighting</modelDescription>
+  <modelName>Philips hue bridge 2015</modelName>
+  <modelNumber>BSB002</modelNumber>
+  <modelURL>http://www.meethue.com</modelURL>
+  <serialNumber>E0CB4EFFFEE16375</serialNumber>
+  <UDN>uuid:2f402f80-da50-11e1-9b23-e0cb4ee16375</UDN>
+  <serviceList>
+  <service>
+  <serviceType>(null)</serviceType>
+  <serviceId>(null)</serviceId>
+  <controlURL>(null)</controlURL>
+  <eventSubURL>(null)</eventSubURL>
+  <SCPDURL>(null)</SCPDURL>
+  </service>
+  </serviceList>
+  <presentationURL>index.html</presentationURL>
+  <iconList>
+  <icon>
+  <mimetype>image/png</mimetype>
+  <height>48</height>
+  <width>48</width>
+  <depth>24</depth>
+  <url>hue_logo_0.png</url>
+  </icon>
+  <icon>
+  <mimetype>image/png</mimetype>
+  <height>120</height>
+  <width>120</width>
+  <depth>24</depth>
+  <url>hue_logo_3.png</url>
+  </icon>
+  </iconList>
+  </device>
+  </root>`
+
+  return(response);
+}
